pyproject.toml
[project]
name = "movie-services"
version = "0.1.0"
description = ""
authors = [
    {name = "Nihal Fateen",email = "nihalahmedfateen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pymongo (>=4.15.5,<5.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "requests (>=2.32.5,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pylint-django (>=2.7.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
